requirements.txt
streamlit
google-play-scraper
app-store-scraper
google-generativeai
supabase
python-dotenv